Mapper 대문자 변환
법인 프로필 업로드 오류 수정
(company-join.html, company-page.html, MyPageController.java)

기존 사용하던
companyMember-mapper.xml 삭제
CompanyMemberService2.java -> CompanyMemberService.java 파일명 변경
기존 CompanyMemberService.java 삭제
CompanyMemberServiceImpl.java 삭제
