+++ src/main/java/com/uniqueGames/controller/JoinController.java
@@ -2,7 +2,7 @@ package com.uniqueGames.controller;
 
 import com.uniqueGames.model.Company;
 import com.uniqueGames.model.Member;
-import com.uniqueGames.service.CompanyMemberService2;
+import com.uniqueGames.service.CompanyMemberService;
 import com.uniqueGames.service.MemberService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,12 +15,12 @@ import org.springframework.web.bind.annotation.*;
 public class JoinController {
 
 	private MemberService memberService;
-	private CompanyMemberService2 companyMemberService2;
+	private CompanyMemberService companyMemberService;
 
 	@Autowired
-	public JoinController(MemberService memberService, CompanyMemberService2 companyMemberService2) {
+	public JoinController(MemberService memberService, CompanyMemberService companyMemberService) {
 		this.memberService = memberService;
-		this.companyMemberService2 = companyMemberService2;
+		this.companyMemberService = companyMemberService;
 	}
 	@GetMapping("join")
 	public String join() {
@@ -46,11 +46,11 @@ public class JoinController {
 	}
 	@PostMapping("joincompany")
 	public String joinCompanyProc(Company company, Model model) {
-		String fileName = companyMemberService2.fileCheck(company.getFile());
+		String fileName = companyMemberService.fileCheck(company.getFile());
 		company.setProfileImg(fileName);
-		int result = companyMemberService2.save(company);
+		int result = companyMemberService.save(company);
 		if(result == 1) {
-			companyMemberService2.fileSave();
+			companyMemberService.fileSave();
 			model.addAttribute("result", "join");
 			model.addAttribute("url", "/login");
 		}else {
