+++ src/main/java/com/uniqueGames/controller/MyPageController.java
@@ -6,7 +6,7 @@ import com.uniqueGames.model.Member;
 import com.uniqueGames.model.SessionConstants;
 import com.uniqueGames.repository.CompanyRepositoryMapper;
 import com.uniqueGames.repository.MemberRepositoryMapper;
-import com.uniqueGames.service.CompanyMemberService2;
+import com.uniqueGames.service.CompanyMemberService;
 import com.uniqueGames.service.MemberService;
 
 import javax.servlet.http.HttpSession;
@@ -18,21 +18,20 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
 
 @Controller
 @Slf4j
 public class MyPageController {
 
     private MemberService memberService;
-    private CompanyMemberService2 companyMemberService2;
+    private CompanyMemberService companyMemberService;
     private MemberRepositoryMapper memberRepositoryMapper;
     private CompanyRepositoryMapper companyRepositoryMapper;
     @Autowired
-    public MyPageController(MemberService memberService, CompanyMemberService2 companyMemberService2,
+    public MyPageController(MemberService memberService, CompanyMemberService companyMemberService,
                             MemberRepositoryMapper memberRepositoryMapper, CompanyRepositoryMapper companyRepositoryMapper) {
         this.memberService = memberService;
-        this.companyMemberService2 = companyMemberService2;
+        this.companyMemberService = companyMemberService;
         this.memberRepositoryMapper = memberRepositoryMapper;
         this.companyRepositoryMapper = companyRepositoryMapper;
     }
@@ -68,7 +67,7 @@ public class MyPageController {
             viewName = "myPage/member-page";
         }else if(mode.contains("Company")) {
             Company company = companyRepositoryMapper.findById(((Company) session.getAttribute(SessionConstants.LOGIN_MEMBER)).getCompanyId());
-            String game = companyMemberService2.gameName(company.getCompanyId());
+            String game = companyMemberService.gameName(company.getCompanyId());
             addr = company.getAddr();
             if(!addr.equals("   ")){
                 String[] addrSplit = addr.split("   ");
@@ -126,23 +125,28 @@ public class MyPageController {
     }
 
     @PostMapping("/companyupdate")
-    public String companyUpdate(HttpSession session, Company company) {
+    public String companyUpdate(HttpSession session, Company company,
+                                @RequestParam("deleteImg") String deleteImg) {
         String oldFile = company.getProfileImg();
-        String fileName = companyMemberService2.fileCheck(company.getFile());
+        String fileName = companyMemberService.fileCheck(company.getFile());
         if(fileName != null) {
             company.setNewProfileImg(fileName);
         }else {
             company.setNewProfileImg(oldFile);
         }
 
-        int result = companyMemberService2.update(company);
+        if(deleteImg.equals("delete")) {
+            companyMemberService.fileDelete(oldFile);
+            company.setNewProfileImg("");
+        }
+
+        int result = companyMemberService.update(company);
         if(result == 1) {
-            companyMemberService2.fileSave();
+            companyMemberService.fileSave();
             if(!oldFile.isEmpty() && !company.getNewProfileImg().equals(oldFile)){
-                companyMemberService2.fileDelete(oldFile);
+                companyMemberService.fileDelete(oldFile);
             }
             session.setAttribute("login", "company");
-            System.out.println("마이페이지 수정 완료");
         }else {
             System.out.println("수정 실패");
         }
@@ -191,7 +195,7 @@ public class MyPageController {
                                 @RequestParam("newpassword") String newpassword) {
         String viewName = "";
         Company company = (Company)session.getAttribute(SessionConstants.LOGIN_MEMBER);
-        int result = companyMemberService2.CmypageNewPass(company.getCompanyId(), password, newpassword);
+        int result = companyMemberService.CmypageNewPass(company.getCompanyId(), password, newpassword);
         if(result == 1) {
             session.invalidate();
             model.addAttribute("result", "change");
