--- src/main/resources/mybatis/companyMember-mapper.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="mapper.companyMember">
-
-	<select id="login" parameterType="Company" resultType="int">
-		select count(*) from company where company_id=#{companyId} and password=#{password}
-	</select>
-
-	<insert id="insert" parameterType="Company">
-		insert into TB_COMPANY (company_id, password, name, email, tel, phone_num, addr)
-		values (#{companyId},#{password},#{name},#{email},#{tel},#{phoneNum},#{addr})
-	</insert>
-	
-	<select id="idCheck" parameterType="java.lang.String" resultType="int">
-		select count(*) from TB_COMPANY where company_id=#{companyId}
-	</select>
-
-	<select id="findId" parameterType="Company" resultType="java.lang.String">
-		select company_id from TB_COMPANY where name=#{name} and phone_num=#{phoneNum}
-	</select>
-
-	<select id="findPwd" parameterType="Company" resultType="int">
-		select count(*) from TB_COMPANY where company_id=#{companyId} and name=#{name} and phone_num=#{phoneNum}
-	</select>
-
-	<update id="update" parameterType="Company">
-		update TB_COMPANY set name=#{name}, email=#{email}, addr=#{addr}, tel=#{tel}, phone_num=#{phoneNum}
-		where company_id=#{companyId} and password=#{password}
-	</update>
-
-	<select id="companyPage" parameterType="java.lang.String" resultType="Company">
-		select company_id, g_id, password, name, email, tel, phone_num, addr from TB_COMPANY where company_id=#{companyId}
-	</select>
-	
-	<update id="changeCpassword" parameterType="Company">
-		update TB_COMPANY set password=#{cnewpassword} where company_id=#{companyId}
-	</update>
-	
-	<select id="getGameName" parameterType="String" resultType="String">
-		select b.name from TB_COMPANY a, TB_GAME b where a.G_ID = b.ID and company_id=#{companyId}
-	</select>
-
-	<delete id="deleteCompany" parameterType="Company">
-		delete from TB_COMPANY where company_id=#{companyId} and password=#{password}
-	</delete>
-
-	<select id="emailCheck" parameterType="java.lang.String" resultType="int">
-		select count(*) from TB_COMPANY where email=#{email}
-	</select>
-	
-	<select id="phoneCheck" parameterType="java.lang.String" resultType="int">
-		select count(*) from TB_COMPANY where phone_num=#{phoneNum}
-	</select>
-
-	<select id="aGetMemberList" resultType="Company">
-		SELECT company_id, name FROM COMPANY
-	</select>
-
-	<select id="aGetDetailMember" parameterType="java.lang.String" resultType="Company">
-		SELECT * FROM COMPANY WHERE COMPANY_ID=#{id}
-	</select>
-</mapper>
