+++ src/main/resources/templates/join/company-join.html
@@ -27,7 +27,6 @@
   <form action="joincompany" name="companyJoin" method="post" id="content-2" enctype="multipart/form-data">
     <input type="hidden" name="type" th:value="join">
     <input type="hidden" name="type2" th:value="company">
-    <input type="hidden" name="profileImg">
     <div>
       <ul>
         <li class="jointab"><a th:href="@{/join}">개인 회원</a></li>
