+++ src/main/resources/templates/myPage/company-page.html
@@ -58,6 +58,7 @@
     <input type="hidden" name="companyId" th:value="${company?.companyId}">
     <input type="hidden" name="profileImg" th:value="${company?.profileImg}">
     <input type="hidden" name="newProfileImg">
+    <input type='hidden' name="deleteImg">
     <div class="div-update">
       <ul>
         <li class="mypage-insert">
